duplicate_similar_image_finder/requirements.txt
imagehash>=4.3
Pillow>=10.0
